Implementação da atualização de gastos

+++ Models/Expense.cs
@@ -5,14 +5,14 @@ namespace cteds_projeto_final.Models
     public class Expense
     {
         public long? expenseId;
-        public float value { get; set;}
+        public decimal value { get; set;}
         public string desc { get; set; }
         public long category_id { get; set; }
         public DateTime? added_dttm { get; set; }
         public DateTime expense_dttm { get; set; }
 
 
-        public Expense(float value, string desc, long category_id, DateTime expense_dttm, DateTime ? added_dttm = null, long? expenseId = null)
+        public Expense(decimal value, string desc, long category_id, DateTime expense_dttm, DateTime ? added_dttm = null, long? expenseId = null)
         {
             this.expenseId = expenseId;
             this.value = value;
